restructured the fetch of all products to be in the Layout component instead of the productsToDisplay, so that it can be used in the search-bar
